- leftover logic functions for db moved from __init__ into db_logic. tests updated imports updated - refs #401
