+++ app/logic/db_logic.py
@@ -8,7 +8,8 @@ import os
 from app.database import db
 from app.profile.models import User, Publisher, PublisherUser, UserRoleEnum
 from app.package.models import BitStore, Package, PackageStateEnum, PackageTag
-
+from app.utils import InvalidUsage
+import app.schemas as schema
 
 def find_or_create_user(user_info, oauth_source='github'):
     """
@@ -34,6 +35,33 @@ def find_or_create_user(user_info, oauth_source='github'):
         db.session.add(user)
         db.session.commit()
     return user
+
+
+def get_user_by_id(user_id):
+    '''
+    Returns user info by given id from DB
+    '''
+    user = User.query.filter_by(id=user_id).first()
+    if not user:
+        raise InvalidUsage("User not found", 404)
+    user_schema = schema.UserSchema()
+    user_info = user_schema.dump(user)
+
+    return user_info.data
+
+
+def get_publisher(publisher):
+    '''
+    Returns publisher info from DB
+    '''
+
+    publisher_info = Publisher.query.filter_by(name=publisher).first()
+    if publisher_info is None:
+        raise InvalidUsage("Publisher not found", 404)
+    publisher_schema = schema.PublisherSchema()
+    info = publisher_schema.dump(publisher_info)
+
+    return info.data
 
 
 def create_or_update_package_tag(publisher_name, package_name, tag):
@@ -151,3 +179,33 @@ def package_exists(publisher_name, package_name):
         .filter(Package.name == package_name,
                 Publisher.name == publisher_name).all()
     return len(instance) > 0
+
+
+def get_metadata_for_package(publisher, package):
+    '''
+    Returns metadata for given package owned by publisher
+    '''
+    data = Package.query.join(Publisher).\
+        filter(Publisher.name == publisher,
+               Package.name == package,
+               Package.status == PackageStateEnum.active).\
+        first()
+    if not data:
+        return None
+
+    metadata_schema = schema.PackageMetadataSchema()
+    metadata = metadata_schema.dump(data).data
+
+    return metadata
+
+
+def get_package_names_for_publisher(publisher):
+    metadata = Package.query.join(Publisher).\
+        with_entities(Package.name).\
+        filter(Publisher.name == publisher).all()
+    if len(metadata) is 0:
+        raise InvalidUsage('No Data Package Found For The Publisher', 404)
+    keys = []
+    for d in metadata:
+        keys.append(d[0])
+    return keys
